% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_and_co2.R
\name{avg_extraction_distance}
\alias{avg_extraction_distance}
\title{Estimate Average Wood Extraction Distance From Forest Road Density}
\usage{
avg_extraction_distance(frd)
}
\arguments{
\item{frd}{Forest road density (truck roads) in m/ha}
}
\value{
The average extraction distance from the felling spot to the nearest
landing at a truck road.
}
\description{
Estimate Average Wood Extraction Distance From Forest Road Density
}
\examples{
  frd <- c(15, 30, 60, 100) # Forest road densities m/ha
  avg_extraction_distance(frd)

}
