% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_and_co2.R
\name{fuel_cons_forwarder_1}
\alias{fuel_cons_forwarder_1}
\title{Fuel Consumption of a Forwarder per Cubic Meter Harvested Wood (Version #1)}
\usage{
fuel_cons_forwarder_1(aed)
}
\arguments{
\item{aed}{Average extraction distance to the nearest truck road}
}
\value{
Fuel consumption of a forwarder in liters diesel fuel per m³ wood to
be handled
}
\description{
Fuel consumption per m³ harvested wood derived from the data provided by
\insertCite{grigolato_et_al_2022;textual}{care4cmodel}. Includes loading,
transportation, and unloading.
}
\examples{
  frd <- c(15, 30, 60, 100) # Forest road densities m/ha
  avg_extraction_distance(frd) |>
    fuel_cons_forwarder_1()

}
\references{
\insertRef{grigolato_et_al_2022}{care4cmodel}
}
