% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_and_co2.R
\name{fuel_cons_road_maintenance}
\alias{fuel_cons_road_maintenance}
\title{Annual Fuel Consumption for Truck Road Network Maintenance}
\usage{
fuel_cons_road_maintenance(frd)
}
\arguments{
\item{frd}{Forest road density in m/ha}
}
\value{
Diesel fuel consumption for truck road network maintenance in
l/ha/year
}
\description{
Estimate the annual diesel fuel consumption per ha for maintaining an
existing truck road network in the forest. Estimate based on
\insertCite{enache_stampfer_2015;textual}{care4cmodel}.
}
\examples{
  frd <- c(15, 30, 60, 100)
  fuel_cons_road_maintenance(frd)

}
\references{
\insertRef{enache_stampfer_2015}{care4cmodel}
}
