% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c4c_co2_result.R
\name{is_c4c_co2_result}
\alias{is_c4c_co2_result}
\title{Check if an Object is of Class c4c_co2_result}
\usage{
is_c4c_co2_result(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
\code{TRUE}, if \code{x} has class c4c_co2_result, \code{FALSE} if
not
}
\description{
Check if an Object is of Class c4c_co2_result
}
\keyword{internal}
