% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c4c_concept.R
\name{is_c4c_concept}
\alias{is_c4c_concept}
\title{Check if an Object is of Class c4c_concept}
\usage{
is_c4c_concept(x)
}
\arguments{
\item{x}{object to check}
}
\value{
\code{TRUE}, if \code{x} has class c4c_concept, \code{FALSE} if not
}
\description{
Check if an Object is of Class c4c_concept
}
\examples{
  data(pine_thinning_from_above_1)
  x <- unclass(pine_thinning_from_above_1)

  is_c4c_concept(pine_thinning_from_above_1)
  is_c4c_concept(x)

}
