% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c4c_co2_result.R
\name{new_c4c_co2_result}
\alias{new_c4c_co2_result}
\title{Constructor for the Class c4c_co2_result}
\usage{
new_c4c_co2_result(x = list())
}
\arguments{
\item{x}{An object to become a \code{c4c_co2_result}}
}
\value{
\code{x}, but with the class attribute \code{c4c_co2_result}
}
\description{
For internal use only. Objects of class c4c_co2_result are only constructed
by functions that evaluate simulation runs. There is no validator function,
for such objects, because their consistency is guaranteed by the functions
that build them.
}
\keyword{internal}
