% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_c4c_concept.R
\name{new_c4c_concept}
\alias{new_c4c_concept}
\title{Constructor for a c4c_concept Object}
\usage{
new_c4c_concept(x = list())
}
\arguments{
\item{x}{a list object}
}
\value{
Returns an object of class \code{c4c_concept}
}
\description{
Constructor for a c4c_concept Object
}
\examples{
  # remove the c4c_class attribute for the example's sake
  x <- unclass(pine_thinning_from_above_1)

  x <- new_c4c_concept(x)

}
