% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pine_no_thinning_and_clearcut_1.R
\docType{data}
\name{pine_no_thinning_and_clearcut_1}
\alias{pine_no_thinning_and_clearcut_1}
\title{pine_no_thinning_and_clearcut_1}
\format{
An object of class \code{c4c_concept} of length 3.
}
\usage{
pine_no_thinning_and_clearcut_1
}
\description{
Generated from simulation runs to mimic a concept of treating a
Scots pine stand with no thinnings and a short final harvest phase with the
model SILVA.
}
\keyword{datasets}
