% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pine_thinning_from_above_1.R
\docType{data}
\name{pine_thinning_from_above_1}
\alias{pine_thinning_from_above_1}
\title{pine_thinning_from_above_1}
\format{
An object of class \code{c4c_concept} of length 3.
}
\usage{
pine_thinning_from_above_1
}
\description{
Generated from simulation runs to mimic a concept similar to the
Scots pine management of the Bavarian State Forest with the model SILVA.
}
\keyword{datasets}
